Done Ui Chat and Filter by Price and Manufacture

+++ src/screens/Compare.js
@@ -24,7 +24,7 @@ const Compare = () => {
   );
   const { CategoryName } = useParams();
   useEffect(() => {
-    if (CategoryName !== categoryName) {
+    if (CategoryName !== null && CategoryName !== categoryName) {
       dispatch(deleteAllCompare());
     }
   }, [categoryName]);
